Small fixes; changed location of collectables, changed energy to float (energy is between 0 and 1)

+++ Assets/Scripts/Jose/EnergyPotion.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class EnergyPotion : Collectable
 {
     //Modifier quantity
-    public int energymod;
+    public float energymod;
 
     public override void Effects(ArturBasePlayerController player)
     {
